client/certabo/requirements.txt
berserk==0.10.0
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
cryptography==3.1.1
Deprecated==1.2.10
idna==2.9
importlib-metadata==1.7.0
jeepney==0.4.3
keyring==21.2.1
lru-dict==1.1.6
ndjson==0.3.1
pycparser==2.20
pyserial==3.4
python-chess==0.31.4
pywin32-ctypes==0.2.0
requests==2.23.0
SecretStorage==3.1.2
six==1.15.0
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
